+++ tests/conftest.py
@@ -4,6 +4,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 import xarray as xr
+
 from calliope.attrdict import AttrDict
 from calliope.backend import latex_backend_model, pyomo_backend_model
 from calliope.preprocess import CalliopeMath
@@ -50,6 +51,13 @@ def simple_supply():
     m = build_model({}, "simple_supply,two_hours,investment_costs")
     m.build()
     m.solve()
+    return m
+
+
+@pytest.fixture
+def simple_supply_build_func():
+    m = build_model({}, "simple_supply,two_hours,investment_costs")
+    m.build()
     return m