+++ tests/test_backend_latex_backend.py
@@ -2,6 +2,7 @@ import textwrap
 
 import pytest
 import xarray as xr
+
 from calliope import exceptions
 from calliope.backend import latex_backend_model
 
@@ -301,6 +302,11 @@ class TestLatexBackendModel:
 
                     foobar
 
+                    \textbf{Uses}:
+                    \begin{itemize}
+                        \item no_dims
+                    \end{itemize}
+
                     \textbf{Default}: 0
 
                     \begin{equation}
@@ -334,6 +340,10 @@ class TestLatexBackendModel:
 
                     foobar
 
+                    **Uses**:
+
+                    * no_dims
+
                     **Default**: 0
 
                     .. container:: scrolling-wrapper
@@ -365,6 +375,10 @@ class TestLatexBackendModel:
                     ### expr
 
                     foobar
+
+                    **Uses**:
+
+                    * [no_dims](#no_dims)
 
                     **Default**: 0
 
@@ -436,7 +450,9 @@ class TestLatexBackendModel:
                     """
         )
 
-    def test_generate_math_doc_mkdocs_tabbed(self, dummy_model_data, dummy_model_math):
+    def test_generate_math_doc_mkdocs_features_tabs(
+        self, dummy_model_data, dummy_model_math
+    ):
         backend_model = latex_backend_model.LatexBackendModel(
             dummy_model_data, dummy_model_math
         )
@@ -448,7 +464,7 @@ class TestLatexBackendModel:
                 "default": 0,
             },
         )
-        doc = backend_model.generate_math_doc(format="md", mkdocs_tabbed=True)
+        doc = backend_model.generate_math_doc(format="md", mkdocs_features=True)
         assert doc == textwrap.dedent(
             r"""
 
@@ -477,18 +493,74 @@ class TestLatexBackendModel:
                     """
         )
 
-    def test_generate_math_doc_mkdocs_tabbed_not_in_md(
+    def test_generate_math_doc_mkdocs_features_admonition(
+        self, dummy_model_data, dummy_model_math
+    ):
+        backend_model = latex_backend_model.LatexBackendModel(
+            dummy_model_data, dummy_model_math
+        )
+        backend_model._add_all_inputs_as_parameters()
+        backend_model.add_global_expression(
+            "expr",
+            {
+                "equations": [{"expression": "no_dims + 1"}],
+                "description": "foobar",
+                "default": 0,
+            },
+        )
+        doc = backend_model.generate_math_doc(format="md", mkdocs_features=True)
+        assert doc == textwrap.dedent(
+            r"""
+
+                    ## Where
+
+                    ### expr
+
+                    foobar
+
+                    ??? info "Uses"
+
+                        * [no_dims](#no_dims)
+
+                    **Default**: 0
+
+                    === "Math"
+
+                        $$
+                        \begin{array}{l}
+                            \quad \textit{no\_dims} + 1\\
+                        \end{array}
+                        $$
+
+                    === "YAML"
+
+                        ```yaml
+                        equations:
+                        - expression: no_dims + 1
+                        ```
+
+                    ## Parameters
+
+                    ### no_dims
+
+                    ??? info "Used in"
+
+                        * [expr](#expr)
+                    """
+        )
+
+    def test_generate_math_doc_mkdocs_features_not_in_md(
         self, dummy_model_data, dummy_model_math
     ):
         backend_model = latex_backend_model.LatexBackendModel(
             dummy_model_data, dummy_model_math
         )
         with pytest.raises(exceptions.ModelError) as excinfo:
-            backend_model.generate_math_doc(format="rst", mkdocs_tabbed=True)
+            backend_model.generate_math_doc(format="rst", mkdocs_features=True)
 
         assert check_error_or_warning(
             excinfo,
-            "Cannot use MKDocs tabs when writing math to a non-Markdown file format.",
+            "Cannot use MKDocs features when writing math to a non-Markdown file format.",
         )
 
     @pytest.mark.parametrize(