+++ tests/test_core_preprocess.py
@@ -1,9 +1,10 @@
 import warnings
 
-import calliope
-import calliope.exceptions as exceptions
 import pandas as pd
 import pytest
+
+import calliope
+import calliope.exceptions as exceptions
 from calliope.attrdict import AttrDict
 
 from .common.util import build_test_model as build_model
@@ -276,7 +277,7 @@ class TestChecks:
         )
 
     def test_unspecified_base_tech(self):
-        """All technologies and technology groups must specify a base_tech"""
+        """All technologies must specify a base_tech"""
         override = AttrDict.from_yaml_string(
             """
             techs.test_supply_no_base_tech:
@@ -292,7 +293,7 @@ class TestChecks:
             build_model(override_dict=override, scenario="simple_supply,one_day")
 
     def test_tech_as_base_tech(self):
-        """All technologies and technology groups must specify a base_tech"""
+        """All technologies must specify a base_tech"""
         override1 = AttrDict.from_yaml_string(
             """
             techs.test_supply_tech_base_tech: