+++ tests/test_preprocess_model_data.py
@@ -5,6 +5,7 @@ import numpy as np
 import pandas as pd
 import pytest
 import xarray as xr
+
 from calliope import exceptions
 from calliope.attrdict import AttrDict
 from calliope.preprocess import data_sources, scenarios
@@ -14,7 +15,7 @@ from .common.util import build_test_model as build_model
 from .common.util import check_error_or_warning
 
 
-@pytest.fixture()
+@pytest.fixture
 def model_def():
     model_def_path = Path(__file__).parent / "common" / "test_model" / "model.yaml"
     model_dict = AttrDict.from_yaml(model_def_path)
@@ -24,7 +25,7 @@ def model_def():
     return model_def_override, model_def_path
 
 
-@pytest.fixture()
+@pytest.fixture
 def data_source_list(model_def, init_config):
     model_def_dict, model_def_path = model_def
     return [
@@ -35,14 +36,14 @@ def data_source_list(model_def, init_config):
     ]
 
 
-@pytest.fixture()
+@pytest.fixture
 def init_config(config_defaults, model_def):
     model_def_dict, _ = model_def
     config_defaults.union(model_def_dict.pop("config"), allow_override=True)
     return config_defaults["init"]
 
 
-@pytest.fixture()
+@pytest.fixture
 def model_data_factory(model_def, init_config, model_defaults):
     model_def_dict, _ = model_def
     return ModelDataFactory(
@@ -50,13 +51,13 @@ def model_data_factory(model_def, init_config, model_defaults):
     )
 
 
-@pytest.fixture()
+@pytest.fixture
 def model_data_factory_w_params(model_data_factory: ModelDataFactory):
     model_data_factory.add_node_tech_data()
     return model_data_factory
 
 
-@pytest.fixture()
+@pytest.fixture
 def my_caplog(caplog):
     caplog.set_level(logging.DEBUG, logger="calliope.preprocess")
     return caplog
@@ -442,7 +443,7 @@ class TestModelData:
             "foo | Cannot pass parameter data as a list unless the parameter is one of the pre-defined lookup arrays",
         )
 
-    def test_inherit_defs_inactive(
+    def test_template_defs_inactive(
         self, my_caplog, model_data_factory: ModelDataFactory
     ):
         def_dict = {"A": {"active": False}}
@@ -452,8 +453,11 @@ class TestModelData:
         assert "(nodes, A) | Deactivated." in my_caplog.text
         assert not new_def_dict
 
-    def test_inherit_defs_nodes_inherit(self, model_data_factory: ModelDataFactory):
-        def_dict = {"A": {"inherit": "init_nodes", "my_param": 1}, "B": {"my_param": 2}}
+    def test_template_defs_nodes_inherit(self, model_data_factory: ModelDataFactory):
+        def_dict = {
+            "A": {"template": "init_nodes", "my_param": 1},
+            "B": {"my_param": 2},
+        }
         new_def_dict = model_data_factory._inherit_defs(
             dim_name="nodes", dim_dict=AttrDict(def_dict)
         )
@@ -467,11 +471,13 @@ class TestModelData:
             "B": {"my_param": 2},
         }
 
-    def test_inherit_defs_nodes_from_base(self, model_data_factory: ModelDataFactory):
+    def test_template_defs_nodes_from_base(self, model_data_factory: ModelDataFactory):
+        """Without a `dim_dict` to start off inheritance chaining, the `dim_name` will be used to find keys."""
         new_def_dict = model_data_factory._inherit_defs(dim_name="nodes")
         assert set(new_def_dict.keys()) == {"a", "b", "c"}
 
-    def test_inherit_defs_techs(self, model_data_factory: ModelDataFactory):
+    def test_template_defs_techs(self, model_data_factory: ModelDataFactory):
+        """`dim_dict` overrides content of base model definition."""
         model_data_factory.model_definition.set_key("techs.foo.base_tech", "supply")
         model_data_factory.model_definition.set_key("techs.foo.my_param", 2)
 
@@ -481,9 +487,10 @@ class TestModelData:
         )
         assert new_def_dict == {"foo": {"my_param": 1, "base_tech": "supply"}}
 
-    def test_inherit_defs_techs_inherit(self, model_data_factory: ModelDataFactory):
+    def test_template_defs_techs_inherit(self, model_data_factory: ModelDataFactory):
+        """Use of template is tracked in updated definition dictionary (as `techs_inheritance` here)."""
         model_data_factory.model_definition.set_key(
-            "techs.foo.inherit", "test_controller"
+            "techs.foo.template", "test_controller"
         )
         model_data_factory.model_definition.set_key("techs.foo.base_tech", "supply")
         model_data_factory.model_definition.set_key("techs.foo.my_param", 2)
@@ -500,7 +507,8 @@ class TestModelData:
             }
         }
 
-    def test_inherit_defs_techs_empty_def(self, model_data_factory: ModelDataFactory):
+    def test_template_defs_techs_empty_def(self, model_data_factory: ModelDataFactory):
+        """An empty `dim_dict` entry can be handled, by returning the model definition for that entry."""
         model_data_factory.model_definition.set_key("techs.foo.base_tech", "supply")
         model_data_factory.model_definition.set_key("techs.foo.my_param", 2)
 
@@ -510,9 +518,10 @@ class TestModelData:
         )
         assert new_def_dict == {"foo": {"my_param": 2, "base_tech": "supply"}}
 
-    def test_inherit_defs_techs_missing_base_def(
+    def test_template_defs_techs_missing_base_def(
         self, model_data_factory: ModelDataFactory
     ):
+        """If inheriting from a template, checks against the schema will still be undertaken."""
         def_dict = {"foo": {"base_tech": "supply"}}
         with pytest.raises(KeyError) as excinfo:
             model_data_factory._inherit_defs(
@@ -528,56 +537,58 @@ class TestModelData:
         [
             ({"my_param": 1}, {"my_param": 1}, None),
             (
-                {"inherit": "foo_group"},
-                {"my_param": 1, "my_other_param": 2, "inherit": "foo_group"},
+                {"template": "foo_group"},
+                {"my_param": 1, "my_other_param": 2, "template": "foo_group"},
                 ["bar_group", "foo_group"],
             ),
             (
-                {"inherit": "bar_group"},
-                {"my_param": 2, "my_other_param": 2, "inherit": "bar_group"},
+                {"template": "bar_group"},
+                {"my_param": 2, "my_other_param": 2, "template": "bar_group"},
                 ["bar_group"],
             ),
             (
-                {"inherit": "bar_group", "my_param": 3, "my_own_param": 1},
+                {"template": "bar_group", "my_param": 3, "my_own_param": 1},
                 {
                     "my_param": 3,
                     "my_other_param": 2,
                     "my_own_param": 1,
-                    "inherit": "bar_group",
+                    "template": "bar_group",
                 },
                 ["bar_group"],
             ),
         ],
     )
-    def test_climb_inheritance_tree(
+    def test_climb_template_tree(
         self,
         model_data_factory: ModelDataFactory,
         node_dict,
         expected_dict,
         expected_inheritance,
     ):
+        """Templates should be found and applied in order of 'ancestry' (newer dict keys replace older ones if they overlap)."""
         group_dict = {
-            "foo_group": {"inherit": "bar_group", "my_param": 1},
+            "foo_group": {"template": "bar_group", "my_param": 1},
             "bar_group": {"my_param": 2, "my_other_param": 2},
         }
-        model_data_factory.model_definition["node_groups"] = AttrDict(group_dict)
-        new_dict, inheritance = model_data_factory._climb_inheritance_tree(
+        model_data_factory.model_definition["templates"] = AttrDict(group_dict)
+        new_dict, inheritance = model_data_factory._climb_template_tree(
             AttrDict(node_dict), "nodes", "A"
         )
         assert new_dict == expected_dict
         assert inheritance == expected_inheritance
 
-    def test_climb_inheritance_tree_missing_ancestor(
+    def test_climb_template_tree_missing_ancestor(
         self, model_data_factory: ModelDataFactory
     ):
+        """Referencing a template that doesn't exist in `templates` raises an error."""
         group_dict = {
-            "foo_group": {"inherit": "bar_group", "my_param": 1},
+            "foo_group": {"template": "bar_group", "my_param": 1},
             "bar_group": {"my_param": 2, "my_other_param": 2},
         }
-        model_data_factory.model_definition["node_groups"] = AttrDict(group_dict)
+        model_data_factory.model_definition["templates"] = AttrDict(group_dict)
         with pytest.raises(KeyError) as excinfo:
-            model_data_factory._climb_inheritance_tree(
-                AttrDict({"inherit": "not_there"}), "nodes", "A"
+            model_data_factory._climb_template_tree(
+                AttrDict({"template": "not_there"}), "nodes", "A"
             )
 
         assert check_error_or_warning(excinfo, "(nodes, A) | Cannot find `not_there`")
@@ -845,7 +856,7 @@ class TestModelData:
 
 
 class TestTopLevelParams:
-    @pytest.fixture()
+    @pytest.fixture
     def run_and_test(self, model_data_factory_w_params):
         def _run_and_test(in_dict, out_dict, dims):
             model_data_factory_w_params.model_definition["parameters"] = {
@@ -1045,7 +1056,7 @@ class TestActiveFalse:
         assert "(nodes, b), (techs, test_storage) | Deactivated" in my_caplog.text
 
     def test_link_active_false(self, my_caplog):
-        overrides = {"tech_groups.test_transmission.active": False}
+        overrides = {"templates.test_transmission.active": False}
         model = build_model(overrides, "simple_storage,two_hours,investment_costs")
 
         # Ensure what should be gone is gone